pyproject.toml
[project]
name = "llm-exp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "outlines>=1.2.8",
    "ollama>=0.6.0",
    "openai>=2.7.1",
    "google-genai",
    "anthropic>=0.72.1",
    "instructor>=1.13.0",
    "seaborn>=0.13.2",
    "langdetect>=1.0.9",
    "tiktoken>=0.12.0",
    "ipywidgets>=8.1.8",
]
